No casty.

+++ src/sdcruntime/object.d
@@ -1,5 +1,5 @@
 module object;
 
-alias int size_t;
+alias long size_t;
 
 extern (C) void* GC_malloc(size_t n);
